--- leetCode/src/leetCode80/leetCode80.java
@@ -1,27 +1,0 @@
-package leetCode80;
-
-/**
- * 
- * @author jason
- * 80. 删除排序数组中的重复项 II
- *
- *
- *
- *
- *
- *
- */
-public class leetCode80 {
-	public static void main(String[] args) {
-		leetCode80 leet80 = new leetCode80();
-		Solution solution = leet80.new Solution();
-		
-	}
-
-	class Solution {
-		public int removeDuplicates(int[] nums) {
-
-			return 0;
-		}
-	}
-}
